Feat Endpoints with authorization

+++ Controllers/CheckInController.cs
@@ -1,3 +1,5 @@
+using System.Security.Claims;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using WorkOutAPI.Data;
 using WorkOutAPI.DTO;
@@ -21,6 +23,7 @@ namespace WorkOutAPI.Controllers
         }
 
         [HttpGet("{userId}")]
+        [Authorize(Roles = "Admin")]
         public async Task<IActionResult> GetListByUserId(int userId, [FromQuery] int page = 1, [FromQuery] int size = 10)
         {
             if(page < 0 || size < 0)
@@ -45,11 +48,18 @@ namespace WorkOutAPI.Controllers
         }
 
         [HttpPost]
+        [Authorize]
         public async Task<IActionResult> Create([FromBody] CheckInCreateDTO model)
         {
             if(!ModelState.IsValid)
             {
                 return BadRequest();
+            }
+
+            bool isOwner = User.FindFirst(ClaimTypes.NameIdentifier)?.Value == model.UserId.ToString();
+            if(!User.IsInRole("Admin") && !isOwner)
+            {
+                return Forbid("You are not authorized");
             }
 
             //To do: Get User by Claim
@@ -73,6 +83,7 @@ namespace WorkOutAPI.Controllers
         }
 
         [HttpDelete("{id}")]
+        [Authorize]
         public async Task<IActionResult> Remove(int id)
         {
             var checkIn = await checkInRepository.GetById(id);
@@ -80,6 +91,12 @@ namespace WorkOutAPI.Controllers
             if(checkIn == null)
             {
                 return NotFound("User not found");
+            }
+
+            bool isOwner = User.FindFirst(ClaimTypes.NameIdentifier)?.Value == checkIn.UserId.ToString();
+            if(!User.IsInRole("Admin") && !isOwner)
+            {
+                return Forbid("You are not authorized");
             }
 
             await checkInRepository.Delete(id);
